requirements.txt
python-telegram-bot==21.1
rapidfuzz==3.10.1
aiosqlite==0.20.0
python-dotenv==1.0.1